moved TraffiCHistory class to its own file
